Make sign up form visible when clicking all sign up buttons from homepage
